backend/requirements.txt
Django==3.2
python-dotenv==1.0.0
djangorestframework==3.12.4
djangorestframework==3.12.4
djoser==2.2.0